[#1] Create a configuration file if one doesn't already exist
A configuration file will be generated on startup if one doesn't already
exist.  This is used to store MySQL connection details and later, other
configuration options.

The file is called config.yml and is stored in the plugin config
directory (/plugins/PowerClaim).  It includes some comments about how to
edit the file.

After the configuration file is created (if necessary), the
configuration options is read and made available to other parts of the
plugin.

Closes #2.