[#1] Save/remove claims from MySQL table
Create a table in the MySQL database to store claimed chunks.  When the
user types /claim, a record is inserted to save the claim.

When a chunk is unclaimed, it is not removed from the database but
instead a field is used to keep track of active claims.  This will allow
for a history function to be implemented later to see who previously
claimed a chunk.

Closes #6.