[#7] Publish JAR file to Maven repository
Upon push, a newly built JAR file will be published to the Maven
repository for this Maven project.

The version number in Gradle includes the branch name - except when
running on the main branch or if it is not known (i.e., a local build).
This allows for a different version to be produced for each branch.

Additionally, the -SNAPSHOT suffix will be used for all non-tag builds
and the -RELEASE suffix for tag builds.